Stop losing half a pointer on 64 bit Windows
The x86_64 build had 9 errors and 37 places where a pointer went through an
int or a long, which on LLP64 keeps 32 bits of 64. All of them are gone and
the build is clean:

  * Menu handles passed as item ids are UINT_PTR, and what ShellExecute() and
    FindExecutable() return is INT_PTR.
  * The 16 dialog procedures return INT_PTR, which is what DLGPROC is on Win64;
    as BOOL they could not even be assigned to the field.
  * _beginthread() returns a uintptr_t, not a ULONG and not a HANDLE.
  * imeheap, a HGLOBAL, is kept in a LONG_PTR.
  * param.c switched on (long)p->var, a pointer holding a small PV_ code.
  * track_vcol() hands a number back through its char * return for two of its
    modes; that goes through intptr_t now, so it survives.

The last of those also covers the numbers that were cast to char_u * and given
to emsg2() for a "%ld": sprintf() then read a pointer sized slot as a long,
which is only the same thing where long is pointer sized. emsgn() takes a long
and there are no such casts left anywhere, which is also the last of the
-Wint-to-pointer-cast warnings on Unix.

Four K&R definitions in termlib.c and two in winjnt.c that only the Windows
build compiles are prototypes now too.

CI compiles both architectures and fails if a truncating cast comes back. The
64 bit build has still never been run, so the release stays 32 bit.

+++ src/param.c
@@ -13,6 +13,8 @@
  * bounds checks to doset().
  */
 
+#include <stdint.h>
+
 #include "vim.h"
 #include "globals.h"
 #include "proto.h"
@@ -1593,7 +1595,7 @@ get_varp(struct param *p)
 	if (!(p->flags & P_IND))
 		return p->var;
 
-	switch ((long)(p->var))
+	switch ((intptr_t)(p->var))
 	{
 		case PV_LIST:	return (char_u *)&(curwin->w_p_list);
 		case PV_NU:		return (char_u *)&(curwin->w_p_nu);
